Move /instance/resolve_dns_cname to a dedicated accoutns api class
